--- operator-pipeline-images/operatorcert/community_prow/main.py
@@ -1,19 +1,0 @@
-"""
-Entrypoint module for community-pipelines prow tests
-"""
-import logging
-from operatorcert.logger import setup_logger
-
-LOGGER = logging.getLogger("operator-cert")
-
-
-def main() -> None:
-    """
-    Main function for initializing prow tests
-    """
-    setup_logger()
-    LOGGER.info("Testing prow")
-
-
-if __name__ == "__main__":  # pragma: no cover
-    main()
