+++ operator-pipeline-images/tests/community_prow/test_main.py
@@ -1,9 +1,3 @@
-from operatorcert.community_prow.main import main
-from unittest.mock import patch, MagicMock
+from operatorcert.entrypoints.community_prow import main
 
-
-# placeholder test for initial entrypoint
-@patch("operatorcert.community_prow.main.setup_logger")
-def test_main(mock_setup_logger: MagicMock) -> None:
-    mock_setup_logger.return_value = "test prow"
-    main()
+main()
